hackathon-caixabank-data-ai-report-1388/requirements.txt
langchain==0.3.3
langchain-ollama==0.2.0
langchain_community
matplotlib==3.9.2
ollama==0.3.3
pandas==2.2.3
numpy>=1.21
requests==2.32.3
seaborn==0.13.2
fpdf2==2.8.1
pytest==8.3.3
scikit-learn==1.2.0
joblib>=1.1.1
statsmodels==0.14.4
fpdf2==2.8.1